Support RegistrationNameAttribute on server type specifying registry key name
Because Microsoft refuses to fix its absurd icon overlay limitation, there is an arms race
under ShellIconOverlayIdentifiers for which handler appears first in the registry.
Currently SharpShell has no armament because it uses the type name for the key name,
where other tools can begin their key names with spaces so that they appear first.
This is especially difficult with Windows 10 which comes installed with 8 overlays
with key names beginning with a space, and it doesn't allow you to delete them.
This change allows specifying the registry key name to use for overlays, and while we're
at it, menu handlers and and other servers which use the type name as a registry key.