feat: more logging options for Property Pages, automatic property page sizing
This commit adds more logging for Property Page extensions. It also
ensures that the size of a SharpPropertySheet is set automatically. This
makes it easier to build a page which has any size in the designer, then
anchor controls to the appropriate location, knowing that the control
will be sized appropriately when shown to the user.