feat: added the IRegistry service
This is the first key step in being able to provide much more extensive
testing for SharpShell. By ensuring all registry access is through an
interface, we can much more easily provide mocks or test
implementations. The end goal is that key operations, such as server
registration, are thoroughly tested to create structures in the registry
of the expected shape.