feat: SharpItemThumbnailHandler for thumbnails for shell items
This base class supports thumbnail handlers initialised with an
IShellItem object.

However,
https://docs.microsoft.com/en-us/windows/win32/api/thumbcache/nn-thumbcache-ithumbnailprovider
suggests that this would only ever be called if a 3rd party API calls
it, not the standard shell. This implementation deliberately makes the
process in-proc, but even then it doesn't seem to be initialised. Looks
like the only valid way to implement IThumbnailHandler is via
IInitializeWithStream.